feat: integrate morphology module
All python and cuda files for morphology operations were integrated to the project. The setup.py was modified to automatically find the .py and .pyx files and create the setuptools Extension() objects. There were some minor bugs with filters/ compilation, so this part was commented out of the setup.py.