feat: add pinned memory + wrapper for operating images by chunks
These are two possible solutions for the problem of not enough memory on GPU to operate on a hole image. The pinned memory didn't solve the problem, for a certain size of image the OS always kills the operation. The chunk solution worked. There may be border issues due to breaking the image in chunks, they must be investigated. The tests were made on the notebook environment only. It still lacks integration with cython and with a multi-GPU environment.