test: organize verification tests for morphology cpu implementations
It is important to have a cpu trustworth implementation to check if cuda operations are correct. Some tests can be verified with opencv, but other functions not implemented in opencv can only be evaluated visualy. Tests coverage is almost complety for cpu implementations, geodesic dilation/erosion and reconstruction by dilation/erosion do not have unit tests, but they are steps to the fill holes algorithm, which has already a test.