refactor: change subtraction oprations to have only one input image
The subtraction operation coudn't be executed by the chunkedExecuter() because it had a second input image, placed in the Args args... part of the chunkedExecuter() wrapper, which the wrapper has no access and pass directly for the destination function. Because of that, the pointer for the second input didn't jump to next chhunk, and the operation could only be correctly performed for the first chunk.