refactor: add template for label images in binary operations
The templated operations for int8 and uint8 will allow fill holes operation on label images of huge data (2048x2048x2048 voxels). Since fill holes is a convergency operation, there is no one of knowing the number of iterations in advance to calculate the amount of padding needed, also, if there are too many operations, there could be no padding size big enough. Because of this, the input image cannot be divided in chunks, the only solution is to reduce the image resolution in order to feat the gpu memory. Information loss is not an issue for label data.