refactor: allow chunkedExecution of geodesic operations + add cython wrapper
Geodesic operations have a mask image, another input that need to have its pointers moved along chunks, also the kernel size is fixed for 8-connectivity, because of that the kernel is not an input. These input differencies made necessary to implement a custum version of the chunkedExecutor for geodesic operations.