test: compare top_hat implementations effect on threashold segmentation
The top_hat_reconstruction is a version of top_hat algorithm used in avizo software. Because it uses morphological reconstruction (a convergency algorithm) it cannot be broken in chunks as the simple top_hat implementation. This implicates that the cuda implementation is limitted to images that fit in the GPU memory, it is a limitation that must be investigated if it is advantageous. For now, it seems it is worth to keep the algorithm despite this momentaneous limatation, while we do not develop for multi-gpu with communication between gpus.