feat: add initial solution for fillHoles in big images
fillHoles couldn't support chunked execution because of the convergency operations. An idea to solve this is to use normal hcunked execution, with margin problems, then repeat the operation with the margins as central columns in the next chunks and do an logical or between bothe results. This way, the mist holes in the margins between chunks would be filled. It isn't fineshed yet. There are problems to assure communication between different gpu's for the logical or operation.