fix: start changing index variables to size_t type
The type int was used in the hole code to keep images size and access indexes. For images with 2000x2000x2000 dimensions, the size and index caused overflow for int type. This problem is resolved chaging all occuracies to size_t type. Only the firt operation was changed for testing, since it worked, the intire code needs refactoring.