refact: change geodesic test allocation of memory
ALlocating three huge arrays on local gpu of the notebook with malloc and than using memset to zero value caused the process to be killed in the local machine. Changing the allocation to use calloc do both taks at once, and allowed the OS to deal with a image of 600x1520x1520 in the local notebook.