Fix: allow chunked execution of morphology functions for huge images
Severak small problems were fixed, the int overflow of indexes, the wronf reading of float test images and memory leakage in smooth operation. Now it is possible to execute all morphology functions (except fill holes) with gpuMemory up to 0.4. Above this limit, it appears an error of invalid input in cudaMemcpy calls to copy data back from the device to the host, in the end of the second chunk execution.