refactor: remove unused pinned versions of morphology operations
These files were created to test two different solutions for operating on huge images: use pinned memory or breaking the image in chunks, and operate over each chunk at a time. The last option was chosen, and the chunkedExecutor() allows to operate on huge images, which do not fit in the gpu memory along with all its copies needed for each algorithm.