tests: add prompt input variables for tests (ngpus, repetitions, verbose)
With this parameters as input values, it is possible to have more flexibility in testing without the need to recompile all the time. Some debugging log outputs were added in the attempt to discover what causes cudamemcpy() calls to fail. It seems that for 8.6GB of data coppied back with cudamemcpy(DeviceToHost), the 'invalid argument' error is hit, only in the second chunk execution.